+++ chat-service/src/main/java/com/ojosama/chatservice/presentation/controller/MessageController.java
@@ -1,18 +1,12 @@
 package com.ojosama.chatservice.presentation.controller;
 
-import com.ojosama.chatservice.application.dto.command.ChangeMessageStatusCommand;
 import com.ojosama.chatservice.application.dto.command.CreateMessageCommand;
 import com.ojosama.chatservice.application.dto.command.DeleteMessageCommand;
-import com.ojosama.chatservice.application.dto.query.FindMessageQuery;
 import com.ojosama.chatservice.application.dto.query.FindMessagesByChatRoomQuery;
 import com.ojosama.chatservice.application.dto.result.MessageResult;
 import com.ojosama.chatservice.application.dto.result.MessageSliceResult;
 import com.ojosama.chatservice.application.service.MessageService;
-import com.ojosama.chatservice.domain.model.MessageStatus;
-import com.ojosama.chatservice.presentation.dto.request.ChangeMessageStatusActionRequest;
-import com.ojosama.chatservice.presentation.dto.request.ChangeMessageStatusRequest;
 import com.ojosama.chatservice.presentation.dto.request.CreateMessageRequest;
-import com.ojosama.chatservice.presentation.dto.response.ChangeMessageStatusResponse;
 import com.ojosama.chatservice.presentation.dto.response.MessageResponse;
 import com.ojosama.chatservice.presentation.dto.response.MessageSliceResponse;
 import com.ojosama.common.response.ApiResponse;
@@ -21,13 +15,13 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -40,22 +34,17 @@ public class MessageController {
     private final MessageService messageService;
 
     @PostMapping("/rooms/{chatRoomId}/messages")
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<ApiResponse<MessageResponse>> createMessage(
             @PathVariable UUID chatRoomId,
-            @RequestHeader("X-User-Id") UUID userId,
+            @AuthenticationPrincipal String userId,
             @Valid @RequestBody CreateMessageRequest request
     ) {
         MessageResult result = messageService.createMessage(
-                new CreateMessageCommand(chatRoomId, userId, request.content())
+                new CreateMessageCommand(chatRoomId, UUID.fromString(userId), request.content())
         );
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(ApiResponse.created(MessageResponse.from(result)));
-    }
-
-    @GetMapping("/messages/{messageId}")
-    public ResponseEntity<ApiResponse<MessageResponse>> getMessage(@PathVariable UUID messageId) {
-        MessageResult result = messageService.getMessage(new FindMessageQuery(messageId));
-        return ResponseEntity.ok(ApiResponse.success(MessageResponse.from(result)));
     }
 
     @GetMapping("/rooms/{chatRoomId}/messages")
@@ -71,31 +60,12 @@ public class MessageController {
     }
 
     @DeleteMapping("/messages/{messageId}")
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<ApiResponse<Void>> deleteMessage(
             @PathVariable UUID messageId,
-            @RequestHeader("X-User-Id") UUID userId
+            @AuthenticationPrincipal String userId
     ) {
-        messageService.deleteMessage(new DeleteMessageCommand(messageId, userId));
+        messageService.deleteMessage(new DeleteMessageCommand(messageId, UUID.fromString(userId)));
         return ResponseEntity.ok(ApiResponse.deleted());
-    }
-
-    @PatchMapping("/messages/{messageId}/status")
-    public ResponseEntity<ApiResponse<ChangeMessageStatusResponse>> changeMessageStatus(
-            @PathVariable UUID messageId,
-            @RequestHeader("X-User-Id") UUID adminId,
-            @Valid @RequestBody ChangeMessageStatusRequest request
-    ) {
-        MessageStatus status = toMessageStatus(request.status());
-        MessageResult result = messageService.changeMessageStatus(
-                new ChangeMessageStatusCommand(messageId, adminId, status)
-        );
-        return ResponseEntity.ok(ApiResponse.success(ChangeMessageStatusResponse.from(result, adminId)));
-    }
-
-    private MessageStatus toMessageStatus(ChangeMessageStatusActionRequest status) {
-        return switch (status) {
-            case ACTIVE -> MessageStatus.ACTIVE;
-            case BLINDED -> MessageStatus.BLINDED;
-        };
     }
 }
